Move: point 도메인 분리를 위해 인프라스트럭쳐 계층 이동
