+++ src/main/kotlin/io/ticketaka/api/common/infrastructure/event/EventDispatcher.kt
@@ -3,8 +3,8 @@ package io.ticketaka.api.common.infrastructure.event
 import io.ticketaka.api.common.domain.DomainEvent
 import io.ticketaka.api.point.domain.PointChargeEvent
 import io.ticketaka.api.point.domain.PointRechargeEvent
-import io.ticketaka.api.reservation.infrastructure.event.PointChargeEventConsumer
-import io.ticketaka.api.reservation.infrastructure.event.PointRechargeEventConsumer
+import io.ticketaka.api.point.infrastructure.event.PointChargeEventConsumer
+import io.ticketaka.api.point.infrastructure.event.PointRechargeEventConsumer
 import org.springframework.stereotype.Component
 
 @Component
