+++ src/main/kotlin/io/ticketaka/api/point/infrastructure/event/PointChargeEventConsumer.kt
@@ -1,36 +1,36 @@
-package io.ticketaka.api.reservation.infrastructure.event
+package io.ticketaka.api.point.infrastructure.event
 
+import io.ticketaka.api.point.domain.PointChargeEvent
 import io.ticketaka.api.point.domain.PointHistory
 import io.ticketaka.api.point.domain.PointHistoryRepository
-import io.ticketaka.api.point.domain.PointRechargeEvent
 import org.springframework.stereotype.Component
 import java.util.concurrent.ConcurrentLinkedQueue
 import kotlin.concurrent.thread
 
 @Component
-class PointRechargeEventConsumer(
+class PointChargeEventConsumer(
     private val pointHistoryRepository: PointHistoryRepository,
 ) {
-    private val eventQueue = ConcurrentLinkedQueue<PointRechargeEvent>()
+    private val eventQueue = ConcurrentLinkedQueue<PointChargeEvent>()
 
     init {
         startEventConsumer()
     }
 
-    private fun consume(events: MutableList<PointRechargeEvent>) {
+    fun consume(events: MutableList<PointChargeEvent>) {
         val pointHistories = mutableListOf<PointHistory>()
         events.forEach { event ->
             PointHistory.newInstance(
                 userId = event.userId,
                 pointId = event.pointId,
                 amount = event.amount,
-                transactionType = PointHistory.TransactionType.RECHARGE,
+                transactionType = PointHistory.TransactionType.CHARGE,
             ).let { pointHistories.add(it) }
         }
         pointHistoryRepository.saveAll(pointHistories)
     }
 
-    fun offer(event: PointRechargeEvent) {
+    fun offer(event: PointChargeEvent) {
         eventQueue.add(event)
     }
 
@@ -38,12 +38,12 @@ class PointRechargeEventConsumer(
         thread(
             start = true,
             isDaemon = true,
-            name = "PointRechargeEventConsumer",
+            name = "PointChargeEventConsumer",
         ) {
             while (true) {
                 if (eventQueue.isNotEmpty()) {
-                    val events = mutableListOf<PointRechargeEvent>()
-                    var quantity = 5
+                    val events = mutableListOf<PointChargeEvent>()
+                    var quantity = 1000
                     while (eventQueue.isNotEmpty().and(quantity > 0)) {
                         quantity--
                         eventQueue.poll()?.let { events.add(it) }
