+++ src/main/kotlin/io/ticketaka/api/point/infrastructure/event/PointRechargeEventConsumer.kt
@@ -1,36 +1,36 @@
-package io.ticketaka.api.reservation.infrastructure.event
+package io.ticketaka.api.point.infrastructure.event
 
-import io.ticketaka.api.point.domain.PointChargeEvent
 import io.ticketaka.api.point.domain.PointHistory
 import io.ticketaka.api.point.domain.PointHistoryRepository
+import io.ticketaka.api.point.domain.PointRechargeEvent
 import org.springframework.stereotype.Component
 import java.util.concurrent.ConcurrentLinkedQueue
 import kotlin.concurrent.thread
 
 @Component
-class PointChargeEventConsumer(
+class PointRechargeEventConsumer(
     private val pointHistoryRepository: PointHistoryRepository,
 ) {
-    private val eventQueue = ConcurrentLinkedQueue<PointChargeEvent>()
+    private val eventQueue = ConcurrentLinkedQueue<PointRechargeEvent>()
 
     init {
         startEventConsumer()
     }
 
-    fun consume(events: MutableList<PointChargeEvent>) {
+    private fun consume(events: MutableList<PointRechargeEvent>) {
         val pointHistories = mutableListOf<PointHistory>()
         events.forEach { event ->
             PointHistory.newInstance(
                 userId = event.userId,
                 pointId = event.pointId,
                 amount = event.amount,
-                transactionType = PointHistory.TransactionType.CHARGE,
+                transactionType = PointHistory.TransactionType.RECHARGE,
             ).let { pointHistories.add(it) }
         }
         pointHistoryRepository.saveAll(pointHistories)
     }
 
-    fun offer(event: PointChargeEvent) {
+    fun offer(event: PointRechargeEvent) {
         eventQueue.add(event)
     }
 
@@ -38,12 +38,12 @@ class PointChargeEventConsumer(
         thread(
             start = true,
             isDaemon = true,
-            name = "PointChargeEventConsumer",
+            name = "PointRechargeEventConsumer",
         ) {
             while (true) {
                 if (eventQueue.isNotEmpty()) {
-                    val events = mutableListOf<PointChargeEvent>()
-                    var quantity = 1000
+                    val events = mutableListOf<PointRechargeEvent>()
+                    var quantity = 5
                     while (eventQueue.isNotEmpty().and(quantity > 0)) {
                         quantity--
                         eventQueue.poll()?.let { events.add(it) }
