+++ src/main/kotlin/io/ticketaka/api/point/infrastructure/jpa/JpaPointHistoryRepository.kt
@@ -1,4 +1,4 @@
-package io.ticketaka.api.reservation.infrastructure.jpa
+package io.ticketaka.api.point.infrastructure.jpa
 
 import io.ticketaka.api.point.domain.PointHistory
 import org.springframework.data.jpa.repository.JpaRepository
